# docs/openAPI.yaml
openapi: 3.0.3
info:
  title: Decentralized S3-Compatible Storage Gateway API
  version: 1.0.0
  description: |
    S3-compatible API for decentralized storage on IPFS with client-side encryption.
    
    This API provides full S3 compatibility while leveraging:
    - IPFS Cluster for distributed persistence
    - Prolly Trees for efficient indexing
    - Client-side HPKE encryption (keys never leave client devices)
    - BLAKE3/Bao for verified streaming
    - CRDTs for conflict-free collaboration
    
    **Authentication**: OAuth 2.0 with JWT Bearer tokens
    
    **Key Features**:
    - Zero-knowledge storage (gateway sees only encrypted blocks)
    - Automatic deduplication via content-addressing
    - Real-time collaboration with CRDTs
    - Versioning and immutable history
    
  contact:
    name: API Support
    url: https://github.com/functionland/dsn-gateway
    email: hi@fx.land
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://gateway.fx.land
    description: Production Gateway
  - url: https://staging.gateway.fx.land
    description: Staging Environment
  - url: http://localhost:9000
    description: Local Development
...
security:
  - oauth2: [storage:read, storage:write]
  - bearerAuth: []
paths:
...
  /:
    get:
      tags:
        - Bucket Operations
      summary: ListBuckets
      operationId: listBuckets
      description: Returns a list of all buckets owned by the authenticated user
      security:
        - oauth2: [storage:read]
      responses:
        '200':
          description: Successfully retrieved bucket list
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ListAllMyBucketsResult'
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <ListAllMyBucketsResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
                  <Owner>
                    <ID>79a59df900b949e55d96a1e698fbacedfd6e09d98eacf8f8d5218e7cd47ef2be</ID>
                    <DisplayName>user@example.com</DisplayName>
                  </Owner>
                  <Buckets>
                    <Bucket>
                      <Name>my-bucket</Name>
                      <CreationDate>2025-12-01T12:00:00.000Z</CreationDate>
                    </Bucket>
                    <Bucket>
                      <Name>photos-2025</Name>
                      <CreationDate>2025-12-05T08:30:00.000Z</CreationDate>
                    </Bucket>
                  </Buckets>
                </ListAllMyBucketsResult>
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /{bucket}/{key+}?uploadId={uploadId}&partNumber={partNumber}:
    parameters:
      - $ref: '#/components/parameters/BucketName'
      - $ref: '#/components/parameters/ObjectKey'
      - name: uploadId
        in: query
        required: true
        description: Upload ID from CreateMultipartUpload
        schema:
          type: string
      - name: partNumber
        in: query
        required: true
        description: Part number (1-10000)
        schema:
          type: integer
          minimum: 1
          maximum: 10000
    
    put:
      tags:
        - Multipart Upload
      summary: UploadPart
      operationId: uploadPart
      description: |
        Uploads a part in a multipart upload. Parts can be uploaded in any order.
        
        **Note**: Each part generates a unique IPFS CID for deduplication.
      security:
        - oauth2: [storage:write]
      parameters:
        - name: Content-Length
          in: header
          required: true
          schema:
            type: integer
            format: int64
            minimum: 5242880  # 5MB minimum except last part
        - name: Content-MD5
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: Part uploaded successfully
          headers:
            ETag:
              description: |
                ETag of the uploaded part (IPFS CID).
                **Important**: Save this for CompleteMultipartUpload.
              schema:
                type: string
              example: '"a54357aff0632cce46d942af68356b38"'
        '400':
          description: Bad Request (part too small, invalid upload ID)
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'
...